+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 1.3.1 (2023-07-03)
+* optimize RwModel
+* rename mut_call to call_mut
+* add macro call_wait and call_mut_wait
+
 ## 1.3.0 (2023-07-03)
 * add RwModel